[refactor] 피커 로직 분리
