+++ app/src/main/java/com/sopt/gongbaek/presentation/ui/auth/component/picker/PickerState.kt
@@ -1,0 +1,9 @@
+package com.sopt.gongbaek.presentation.ui.auth.component.picker
+
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
+
+class PickerState {
+    var selectedItem by mutableStateOf("")
+}
